requirements.txt
pyfiglet~=0.8.post1
RPi.GPIO